YL:W-3318 (NotFoundException): consolidate the notfounds into one exception class
