+++ MusicSchool/Exceptions/DeleteEntityConflict.cs
@@ -4,8 +4,8 @@ namespace MusicSchool.Exceptions;
 
 public class DeleteEntityConflict : BaseException
 {
-    public DeleteEntityConflict()
-        : base("Unable to delete", HttpStatusCode.Conflict)
+    public DeleteEntityConflict(string message)
+        : base(message, HttpStatusCode.Conflict)
     {
     }
 }
