--- MusicSchool/Exceptions/InstrumentNotFoundException.cs
@@ -1,15 +1,0 @@
-﻿using System.Net;
-
-namespace MusicSchool.Exceptions;
-
-public class InstrumentNotFoundException : BaseException
-{
-    public InstrumentNotFoundException(int id)
-        : base($"Instrument with id {id} not found", HttpStatusCode.NotFound)
-    {
-    }
-    public InstrumentNotFoundException(string id)
-    : base($"Instrument with id {id}, not found", HttpStatusCode.NotFound)
-    {
-    }
-}
