+++ MusicSchool/Exceptions/NotFoundException.cs
@@ -1,0 +1,11 @@
+﻿using System.Net;
+
+namespace MusicSchool.Exceptions;
+
+public class NotFoundException : BaseException
+{
+    public NotFoundException(string message)
+        : base(message, HttpStatusCode.NotFound)
+    {
+    }
+}
