--- MusicSchool/Exceptions/StudentNotFoundException.cs
@@ -1,11 +1,0 @@
-﻿using System.Net;
-
-namespace MusicSchool.Exceptions;
-
-public class StudentNotFoundException : BaseException
-{
-    public StudentNotFoundException(int id)
-        : base($"Student with id {id} not found", HttpStatusCode.NotFound)
-    {
-    }
-}
